:bug: Fix: no-cycle에러 해결
