+++ src/components/Diary/ChallengeSection/Challenge/Challenge.js
@@ -2,7 +2,7 @@ import './Challenge.scss';
 
 import AddChallenge from 'components/Diary/Modal/AddChallenge';
 import ChallengeList from 'components/Diary/Modal/ChallengeList';
-import { useModal } from 'components/common/Modal/context/useModal';
+import { useModal } from 'components/common/Modal/context/ModalContext';
 
 import AchievementRate from './AchievementRate/AchievementRate';
 import GoalsList from './Goals/GoalsList';
